pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "brewing"
version = "0.1.0"
description = "Layer-wise Mechanistic Interpretability Framework"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
model = [
    "torch>=2.0",
    "transformers>=4.35",
    "nnsight>=0.2",
]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["brewing*"]
